Rework bed to subscribe to cutscene finished events
