+++ Assets/ScriptableObjects/Cutscenes/LoriTutPart4PlayerGoBedCutscene.asset
@@ -12,7 +12,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d3a435fc937754f479d7b9a96b05cf4f, type: 3}
   m_Name: LoriTutPart4PlayerGoBedCutscene
   m_EditorClassIdentifier: 
-  <Name>k__BackingField: Lori Tut Part 4
+  <Name>k__BackingField: Lori Tut Part 4 Player Go Bed
   <PlayableAsset>k__BackingField: {fileID: 11400000, guid: 8d0f3e0fbb401d140866ca91852f2898,
     type: 2}
   <ConversationSettings>k__BackingField:
