+++ Assets/Scripts/Environment/Office/Bed.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c8a133b0d05a95546b943c0f6e0d56f6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
